Adds keys export/import coverage
