+++ core/services/keystore/keys/p2pkey/export_test.go
@@ -1,4 +1,4 @@
-package csakey
+package p2pkey
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 	"github.com/smartcontractkit/chainlink/core/services/keystore/keys"
 )
 
-func TestCSAKeys_ExportImport(t *testing.T) {
+func TestP2PKeys_ExportImport(t *testing.T) {
 	keys.RunKeyExportImportTestcase(t, createKey, decryptKey)
 }
