+++ core/services/keystore/keys/terrakey/export_test.go
@@ -1,4 +1,4 @@
-package csakey
+package terrakey
 
 import (
 	"testing"
@@ -6,12 +6,12 @@ import (
 	"github.com/smartcontractkit/chainlink/core/services/keystore/keys"
 )
 
-func TestCSAKeys_ExportImport(t *testing.T) {
+func TestTerraKeys_ExportImport(t *testing.T) {
 	keys.RunKeyExportImportTestcase(t, createKey, decryptKey)
 }
 
 func createKey() (keys.KeyType, error) {
-	return NewV2()
+	return New(), nil
 }
 
 func decryptKey(keyJSON []byte, password string) (keys.KeyType, error) {
